Update code improve from comment Card.js
